Added 'B' command to request browse.
This command broadcasts an MDNS request for all
members of a given service type (e.g. _raop._tcp)
to announce their presence.

So far, when I try this, I also immediately get all
the information I need to resolve each service,
but I don't know if this is always going to be the case.