Corrected serialization of domain names.
Domain name encoding is not the same as string encoding
used in TXT records, even when the domain name is not
compressed. I learned that the domain name is split into
labels delimited by '.'.  Each label is independently
prefixed with a length byte.

I found this by creating a unit test to make sure I
could serialize and deserialize an A record, and found
that it was failing. Debugging led me to realize I have
to fix the domain name serializer before doing anything else.