Simplified RR constructor.
All our records have Class.IN (Internet), and all record types
are determined by whichever class derives from the base class Record.
Therefore, it is verbose to pass in Class, and error-prone to
pass in Type. I refactored the RR constructor so neither needs
to be passed in.

Added an abstract function Recorfd.RecordType() to return the Type for
each derived class. RR calls this function so that the Type always
matches the record object's type.

Eliminated the last remaining usage of ReadUInt16(-2) to determine
a binary record's length. Instead, any time a constructor needs to know
the length to deserialized, it is just passed in.